+++ KOF.Bot/pch.h
@@ -65,5 +65,6 @@ extern bool ConsoleCommand(const std::string & input, std::string & out);
 
 extern bool KillProcessesByFileName(const char* fileName);
 
-extern std::string GenerateUniqueString(int iLength);
+extern std::string GenerateUniqueString(size_t iLength);
+
 #endif //PCH_H
